LockPicking refactor
