+++ Plugin/Controllers/BearRifleBehaviour.cs
@@ -4,7 +4,7 @@ using EFT.InventoryLogic;
 using SkillsExtended.Helpers;
 using SkillsExtended.Models;
 using System.Collections.Generic;
-using JetBrains.Annotations;
+using SkillsExtended.Skills;
 using UnityEngine;
 
 namespace SkillsExtended.Controllers;
