+++ Plugin/Skills/WeaponClasses.cs
@@ -1,0 +1,8 @@
+﻿namespace SkillsExtended.Skills;
+
+public struct OrigWeaponValues
+{
+    public float ergo;
+    public float weaponUp;
+    public float weaponBack;
+}
