+++ Server/locale/en.json
@@ -8,13 +8,13 @@
     "FieldMedicineDescription": "Field surgery expertise facilitates application of splints and bandages.",
     "FieldMedicineLevelingUpDescription": "The Field Medicine skill is improved by treating fractures or bleeds with splints and bandages.",
     "FieldMedicineSpeed": "Increases speed of splinting and bandaging by [{0:0.#%}]",
-    "UsecArsystems": "USEC Rifle and Carbine Systems",
-    "UsecArsystemsDescription": "As a USEC PMC, you excel in the use of NATO assault rifles and carbines.",
+    "UsecArsystems": "NATO Weapon Proficiency",
+    "UsecArsystemsDescription": "Proficiency in the use of NATO weapon systems.",
     "UsecArsystemsLevelingUpDescription": "Improved by hitting targets with NATO weapon systems.",
     "UsecArSystemsErgo": "Increases the ergonomics of NATO weapons by [{0:0.#%}]",
     "UsecArSystemsRecoil": "Reduces the recoil of NATO weapons by [{0:0.#%}]",
-    "BearAksystems": "BEAR Rifle and Carbine Systems",
-    "BearAksystemsDescription": "As a BEAR PMC, you excel in the use of Russian assault rifles and carbines.",
+    "BearAksystems": "Eastern Weapon Proficiency",
+    "BearAksystemsDescription": "Proficiency in the use of Eastern weapon systems.",
     "BearAksystemsLevelingUpDescription": "Improved by hitting targets with Russian weapon systems.",
     "BearAkSystemsErgo": "Increases the ergonomics of Russian weapons by [{0:0.#%}]",
     "BearAkSystemsRecoil": "Reduces the recoil of Russian weapons by [{0:0.#%}]",
